BioSequenceMappings

Documentation for BioSequenceMappings.

BioSequenceMappings.Alignment

BioSequenceMappings.Alignment

BioSequenceMappings.Alignment

BioSequenceMappings.Alphabet

BioSequenceMappings.compute_mapping

BioSequenceMappings.compute_weights

BioSequenceMappings.compute_weights!

BioSequenceMappings.default_alphabet

BioSequenceMappings.pairwise_hamming

BioSequenceMappings.pairwise_hamming

BioSequenceMappings.read_fasta

BioSequenceMappings.site_specific_frequencies

BioSequenceMappings.subsample

BioSequenceMappings.subsample_random

BioSequenceMappings.Alignment — Type

mutable struct Alignment{T} where T <: Integer

    dat::Matrix{T}
    alphabet::Union{Nothing, Alphabet{T}}
    weights::Vector{Float64} = ones(size(dat,1))/size(dat,1) # phylogenetic weights of sequences
    names::Vector{String} = fill("", size(dat, 1))

Biological sequences as vectors of type T<:Integer. dat stores sequences in columns: size(dat) returns a tuple (L, M) with L the length and M the number of sequences.

alphabet represents the mapping between integers in dat and biological symbols (nucleotides, amino acids...). If nothing, the alignment cannot be mapped to biological sequences.

weights represent phylogenetic weights, and are initialized to 1/M. They must sum to 1. names are the label of sequences, and are expected to be in the same order as the columns of dat. They do not have to be unique, and can be ignored

Important: When built from a matrix, will transpose the input; if size(dat) = (M, L),

X=Alignment(dat, alphabet) will return an object with size(X.dat) = (L, M).

In other words, assumes that the input matrix has sequences as rows.

Methods

getindex(X::Alignment, i) returns a matrix/vector X.dat[:, i].

for s in X::Alignment iterates over sequences.

eachsequence(X::Alignment) returns an iterator over sequences (Vector{Int}).

eachsequence_weighted(X::Alignment) returns an iterator over sequences and weights as tuples

subaln(X::Alignment, idx) constructs the subaln defined by index idx.

BioSequenceMappings.Alignment — Method

Alignment(data::AbstractMatrix{T}; alphabet = :auto, kwargs...)

Keyword argument alphabet can be :auto, :none or nothing, or an input to Alphabet. Other keyword arguments are passed to the default constructor of Alignment.

BioSequenceMappings.Alignment — Method

Alignment(data::AbstractMatrix, alphabet; kwargs...)

data is a matrix of integers. alphabet can be either - an Alphabet - nothing - something to build an alphabet from: the constructor Alphabet will be called. (e.g. a symbol like :aa, a string, ...)

If the types of alphabet and data mismatch, data is converted.

BioSequenceMappings.Alphabet — Type

struct Alphabet{T}
    string::String
    char_to_index::Dict{Char, T}
    index_to_char::Dict{T, Char}
end

Alphabet can be constructed

from a String and an optional type: Alphabet(string::String[, T<:Integer])

from a mapping Dict{Char, T} where T<:Integer: Alphabet(mapping)

from a Symbol, using default alphabets

from an integer, using default alphabets (see default_alphabets).

BioSequenceMappings.compute_mapping — Method

compute_mapping(s::AbstractString)

Return a Dict{Int, Char}: Dict(i => c for (i,c) in enumerate(s)).

BioSequenceMappings.compute_weights — Function

compute_weights(X::AbstractAlignment, θ = 0.2; normalize = true)

Compute phylogenetic correction weights for sequences of X. The weight sequence S is 1/N, where N is the number of sequences in X at hamming distance less than H from S (including S itself). The threshold H is floor(θ⋅L) where L is the sequence length.

The return value is a tuple (weights, Meff), where Meff is the sum of weights (pre-normalization). If normalize, weights are normalized to sum to one. .

BioSequenceMappings.compute_weights! — Method

compute_weights!(X, θ; kwargs...)

Compute and set weights for X. See compute_weights.

BioSequenceMappings.default_alphabet — Method

default_alphabet(q::Int, T::Type)

if q==21, amino acids

if q==5, nucleotides

if q==4, nucleotides without gaps

if q==2, binary (0, 1)

else, if q<21, return the restriction of amino acids to the first q sites

if q>21, fails

BioSequenceMappings.pairwise_hamming — Method

pairwise_hamming(X, Y; step=1, step_left, step_right)

Return matrix of all hamming distances between sequences of X and Y. Only consider sequences every step. step_left and step_right can be used to skip sequence either in X or in Y:

BioSequenceMappings.pairwise_hamming — Method

pairwise_hamming(X; step, as_vec=true)

Vector of pairwise hamming distances of sequences in X, ordered as [H(1,2), H(1,3), ..., H(M-1, M)] with H standing for hamming distance. If as_vec=false, will return a Matrix instead.

BioSequenceMappings.read_fasta — Method

read_fasta(fastafile::AbstractString; alphabet = :auto, kwargs...)
read_fasta(
    fastafile::AbstractString, alphabet;
    weights = false, theta = 0.2, verbose = false,
)

BioSequenceMappings.site_specific_frequencies — Method

site_specific_frequencies(X::AbstractAlignment, weights=nothing)

Return the site specific frequencies of X. If as_vec (default), the result is a vector of length Lxq. Otherwise, it is a matrix of q rows and L columns

BioSequenceMappings.subsample — Method

subsample(X::AbstractAlignment, indices)

Return an Alignment containing only the sequences of X at indices.

BioSequenceMappings.subsample_random — Method

subsample_random(X::AbstractAlignment, m::Int)

Return an Alignment with m sequences taking randomly from X. Sampling is done without replacement, meaning the m sequences are all at different positions in X.